removed parsing from client

+++ client/parsing.h
@@ -28,7 +28,7 @@ void parseIPArg(char *ip) {
 void parsePortArg(char *port) {
   for (int i = 0; i < strlen(port); i++) {
     if (port[i] != '0') {
-      int port_parsed = strtol(port, NULL, 10);
+      long port_parsed = strtol(port, NULL, 10);
       if (port_parsed <= 0 || port_parsed > 65535) {
         fprintf(stderr, MSG_INVALID_PORT);
         exit(EXIT_FAILURE);
